fix(ci): drop a redundant Windows cast, install libxdo for the database job
- `window.hwnd()` already returns `isize` on Windows, so the `as isize` was an
  `unnecessary_cast` under `-D warnings`. Only the Windows job could see it.
- Making the app menu cross-platform means `muda` — and therefore `libxdo` — is
  now linked on Linux. The MySQL/Postgres job installs the webview deps but not
  that one, so it failed at link time with `unable to find library -lxdo`.

Since this is a new system requirement for every Linux consumer, it's documented
in the getting-started prerequisites alongside the WebKitGTK/GTK packages.

+++ framework/src/shell.rs
@@ -471,7 +471,7 @@ fn window_menu(
         use tao::platform::windows::WindowExtWindows;
         // SAFETY: the HWND comes from the window we were handed and outlives the
         // menu, which is kept alive by the caller.
-        let _ = unsafe { menu.init_for_hwnd(window.hwnd() as isize) };
+        let _ = unsafe { menu.init_for_hwnd(window.hwnd()) };
     }
     #[cfg(target_os = "linux")]
     {
